+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGE LOG
 ## V0.2 Alpha (Upcoming)
 
 * Upgraded to laravel 4.1+
+* Updated to my new cloudflare api package
 * Minor refactoring